Method to get an array of R values from a sensor

Takes in a sensor object and returns a vector of valid R values

# S4 method for Sensor
getR(object, by = 0.01, edgeBy = 100)

Arguments

object
A sensor object

by
The "by" argument for seq. Specifies intervals between subsequent R values

edgeBy
A multiplier for the precision of values near the edges In the range (Rmin, Rmin + by) and (Rmax - by, Rmax), the by argument will be divided by edgeBy.

Value

A numeric array between Rmin and Rmax

Examples

my_sensor <- new("Sensor", Rmin = 1, Rmax = 5, delta = 0.5)
getR(my_sensor, by = 0.1)

#>   [1] 1.000 1.001 1.002 1.003 1.004 1.005 1.006 1.007 1.008 1.009 1.010 1.011
#>  [13] 1.012 1.013 1.014 1.015 1.016 1.017 1.018 1.019 1.020 1.021 1.022 1.023
#>  [25] 1.024 1.025 1.026 1.027 1.028 1.029 1.030 1.031 1.032 1.033 1.034 1.035
#>  [37] 1.036 1.037 1.038 1.039 1.040 1.041 1.042 1.043 1.044 1.045 1.046 1.047
#>  [49] 1.048 1.049 1.050 1.051 1.052 1.053 1.054 1.055 1.056 1.057 1.058 1.059
#>  [61] 1.060 1.061 1.062 1.063 1.064 1.065 1.066 1.067 1.068 1.069 1.070 1.071
#>  [73] 1.072 1.073 1.074 1.075 1.076 1.077 1.078 1.079 1.080 1.081 1.082 1.083
#>  [85] 1.084 1.085 1.086 1.087 1.088 1.089 1.090 1.091 1.092 1.093 1.094 1.095
#>  [97] 1.096 1.097 1.098 1.099 1.100 1.200 1.300 1.400 1.500 1.600 1.700 1.800
#> [109] 1.900 2.000 2.100 2.200 2.300 2.400 2.500 2.600 2.700 2.800 2.900 3.000
#> [121] 3.100 3.200 3.300 3.400 3.500 3.600 3.700 3.800 3.900 4.000 4.100 4.200
#> [133] 4.300 4.400 4.500 4.600 4.700 4.800 4.900 4.901 4.902 4.903 4.904 4.905
#> [145] 4.906 4.907 4.908 4.909 4.910 4.911 4.912 4.913 4.914 4.915 4.916 4.917
#> [157] 4.918 4.919 4.920 4.921 4.922 4.923 4.924 4.925 4.926 4.927 4.928 4.929
#> [169] 4.930 4.931 4.932 4.933 4.934 4.935 4.936 4.937 4.938 4.939 4.940 4.941
#> [181] 4.942 4.943 4.944 4.945 4.946 4.947 4.948 4.949 4.950 4.951 4.952 4.953
#> [193] 4.954 4.955 4.956 4.957 4.958 4.959 4.960 4.961 4.962 4.963 4.964 4.965
#> [205] 4.966 4.967 4.968 4.969 4.970 4.971 4.972 4.973 4.974 4.975 4.976 4.977
#> [217] 4.978 4.979 4.980 4.981 4.982 4.983 4.984 4.985 4.986 4.987 4.988 4.989
#> [229] 4.990 4.991 4.992 4.993 4.994 4.995 4.996 4.997 4.998 4.999 5.000